Move dynamic loader into common
