+++ taichi/rhi/cuda/cuda_driver.h
@@ -2,7 +2,7 @@
 
 #include <mutex>
 
-#include "taichi/system/dynamic_loader.h"
+#include "taichi/common/dynamic_loader.h"
 #include "taichi/common/core.h"
 #include "taichi/rhi/cuda/cuda_types.h"
